# contracts/openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      description: Stable error envelope returned by every ULOO Core endpoint.
      properties:
        code:
          description: Stable machine-readable error code
          title: Code
          type: string
        details:
          anyOf:
          - {}
          - type: 'null'
          title: Details
        message:
          description: Safe human-readable error message
          title: Message
          type: string
        request_id:
          title: Request Id
          type: string
        trace_id:
          title: Trace Id
          type: string
      required:
      - code
      - message
      - request_id
      - trace_id
      title: ErrorResponse
      type: object
...
    TaskCreate:
      properties:
        query:
          description: User's original request
          maxLength: 100000
          minLength: 1
          title: Query
          type: string
        team_id:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional explicit team selection
          title: Team Id
        title:
          description: Short task title
          maxLength: 512
          minLength: 1
          title: Title
          type: string
      required:
      - title
      - query
      title: TaskCreate
      type: object
    TaskDetailResponse:
      description: Task plus its current plan and latest run summary.
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        current_plan_version:
          title: Current Plan Version
          type: integer
        id:
          title: Id
          type: string
        plans:
          items:
            $ref: '#/components/schemas/TaskPlanResponse'
          title: Plans
          type: array
        query:
          title: Query
          type: string
        session_id:
          title: Session Id
          type: string
        status:
          title: Status
          type: string
        team_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Team Id
        title:
          title: Title
          type: string
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
      required:
      - id
      - title
      - query
      - status
      - session_id
      - current_plan_version
      title: TaskDetailResponse
      type: object
    TaskListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TaskResponse'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - offset
      - limit
      title: TaskListResponse
      type: object
    TaskPlanRequest:
      properties:
        feedback:
          anyOf:
          - maxLength: 4000
            type: string
          - type: 'null'
          title: Feedback
      title: TaskPlanRequest
      type: object
    TaskPlanResponse:
      properties:
        approval_status:
          default: pending
          pattern: pending|approved|rejected
          title: Approval Status
          type: string
        assumptions:
          items:
            type: string
          title: Assumptions
          type: array
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        id:
          title: Id
          type: string
        is_mock:
          default: false
          title: Is Mock
          type: boolean
        planner_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Planner Run Id
        questions:
          items:
            type: string
          title: Questions
          type: array
        recommended_team_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Recommended Team Id
        rejection_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejection Reason
        runtime_type:
          default: agno
          title: Runtime Type
          type: string
        steps:
          items:
            $ref: '#/components/schemas/TaskPlanStep'
          title: Steps
          type: array
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        task_id:
          title: Task Id
          type: string
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
        usage:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Usage
        version:
          title: Version
          type: integer
      required:
      - id
      - task_id
      - version
      title: TaskPlanResponse
      type: object
    TaskPlanStep:
      description: One step of an execution plan.
      properties:
        assigned_agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned Agent Id
        depends_on:
          items:
            type: string
          title: Depends On
          type: array
        description:
          anyOf:
          - maxLength: 2000
            type: string
          - type: 'null'
          title: Description
        expected_output:
          anyOf:
          - type: string
          - type: 'null'
          title: Expected Output
        id:
          description: Stable step id within the plan
          title: Id
          type: string
        status:
          default: pending
          enum:
          - pending
          - running
          - completed
          - failed
          - skipped
          title: Status
          type: string
        title:
          maxLength: 256
          minLength: 1
          title: Title
          type: string
      required:
      - id
      - title
      title: TaskPlanStep
      type: object
    TaskResponse:
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        current_plan_version:
          title: Current Plan Version
          type: integer
        id:
          title: Id
          type: string
        query:
          title: Query
          type: string
        session_id:
          title: Session Id
          type: string
        status:
          title: Status
          type: string
        team_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Team Id
        title:
          title: Title
          type: string
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
      required:
      - id
      - title
      - query
      - status
      - session_id
      - current_plan_version
      title: TaskResponse
      type: object
...
info:
  title: ULOO Core
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/tasks:
    get:
      description: List tasks in the current workspace, newest first.
      operationId: list_tasks_api_v1_tasks_get
      parameters:
      - description: Search in title and query
        in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search in title and query
          title: Q
      - description: Filter by lifecycle status
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by lifecycle status
          title: Status
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: List Tasks
      tags:
      - tasks
    post:
      description: Create a task and persist the user's original request.
      operationId: create_task_api_v1_tasks_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetailResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Create Task
      tags:
      - tasks
  /api/v1/tasks/{task_id}:
    get:
      description: Return a task with its plans (plans appear once the Planner has produced them).
      operationId: get_task_api_v1_tasks__task_id__get
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          format: uuid
          title: Task Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetailResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Get Task
      tags:
      - tasks
...
  /api/v1/tasks/{task_id}/plan:
    post:
      description: Use a real structured Agno Planner run to create the next plan version.
      operationId: plan_task_api_v1_tasks__task_id__plan_post
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          format: uuid
          title: Task Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TaskPlanRequest'
              - type: 'null'
              title: Body
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPlanResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Plan Task
      tags:
      - tasks
...